putting the dragging script in it's own file
