+++ src/index.js
@@ -2,6 +2,7 @@ import 'material-icons/iconfont/material-icons.css';
 import './style.css';
 
 import Tasks from './class';
+import handleDrag from './dragging';
 
 const listParent = document.querySelector('.list');
 const clearAll = document.querySelector('.clear');
@@ -112,58 +113,4 @@ clearAll.addEventListener('click', () => {
   render();
 });
 
-///////////////////////////////////////
-
-function getDragAfterElm(y) {
-  const elements = [...document.querySelectorAll('li:not(.dragging)')];
-
-  return elements.reduce(
-    (closest, child) => {
-      const box = child.getBoundingClientRect();
-      const offset = y - box.top - box.height / 2;
-
-      if (offset < 0 && offset > closest.offset) {
-        return { offset, element: child };
-      } else {
-        return closest;
-      }
-    },
-    { offset: Number.NEGATIVE_INFINITY }
-  ).element;
-}
-
-document.querySelectorAll('li').forEach((li) => {
-  li.addEventListener('dragstart', (e) => {
-    e.target.classList.add('dragging');
-  });
-
-  li.addEventListener('dragend', () => {
-    document
-      .querySelectorAll('li')
-      .forEach((x) => x.classList.remove('dragging'));
-  });
-
-  li.addEventListener('dragover', (e) => {
-    e.preventDefault();
-  });
-
-  li.addEventListener('drop', (e) => {
-    e.preventDefault();
-    const draggable = document.querySelector('.dragging');
-    const afterElm = getDragAfterElm(e.clientY);
-
-    if (afterElm === undefined) {
-      listParent.append(draggable);
-    } else {
-      listParent.insertBefore(draggable, afterElm);
-    }
-
-    document.querySelectorAll('li').forEach((elm, i) => {
-      const id = Number(elm.id.split('-')[1]);
-
-      tasks.sort(id, i + 1);
-    });
-
-    render();
-  });
-});
+handleDrag(tasks);
